fixed errors from previous push, improved certain functionalities, such as move course to antoher semester and the course search
