+++ src/Planner.tsx
@@ -1,210 +1,104 @@
+/* eslint-disable indent */
 /* eslint-disable no-extra-parens */
-import React, { useState } from "react";
+import React from "react";
 import "./Planner.css";
 import CourseList from "./CourseList";
-import catalog from "./catalog.json";
-
-interface Course {
-    code: string;
-    name: string;
-    descr: string;
-    credits: string;
-    preReq: string;
-    restrict: string;
-    breadth: string;
-    typ: string;
-}
-
-interface Semester {
-    id: string;
-    season: string;
-    year: string;
-    courses: Course[];
-    skip: boolean;
-}
-
-interface Plan {
-    id: string;
-    title: string;
-    semesters: Semester[];
-}
+import { Plan } from "./interfaces/Plan";
+import { usePlannerFunctions } from "./PlannerFunctions";
 
 interface Planner {
     plan: Plan;
 }
 
 const Planner: React.FC<Planner> = ({ plan }) => {
-    const [semestersData, setSemestersData] = useState(plan.semesters);
-    const [searchTerm, setSearchTerm] = useState("");
-    const [showCourseList, setShowCourseList] = useState(false);
-    const [coursesVisibility, setCoursesVisibility] = useState(
-        Array(semestersData.length).fill(true)
-    );
-    const [allSemestersVisible, setAllSemestersVisible] = useState(true);
+    const {
+        semestersData,
+        searchTerm,
+        showCourseList,
+        coursesVisibility,
+        allSemestersVisible,
+        selectedSemester,
+        setSelectedSemester,
+        handleSearchTermChange,
+        filteredCourses,
+        semesterOptions,
+        handleCourseSelected,
+        handleSkipToggle,
+        handleInsertCourse,
+        handleInsertSemester,
+        handleEditCourse,
+        handleClearCourses,
+        handleRemoveCourse,
+        handleRemoveSemester,
+        handleClearAllSemesters,
+        handleToggleVisibility,
+        handleToggleAllSemestersVisibility,
+        handleMoveCourseStart,
+        moveCourseData,
+        handleMoveCourseConfirm,
+        setMoveCourseData,
+        plans,
+        currentPlanIndex,
+        handlePlanChange,
+        handleInsertPlan,
+        handleRemovePlan
+    } = usePlannerFunctions(plan);
 
-    const filteredCourses = Object.values(catalog)
-        .flatMap(department => Object.values(department))
-        .filter(course => {
-            const normalizedSearchTerm = searchTerm.toLowerCase();
-            const matchesSearchTerm =
-                course.code.toLowerCase().includes(normalizedSearchTerm) ||
-                course.code.split(" ")[1].includes(normalizedSearchTerm);
-
-            return matchesSearchTerm;
-        });
-    const handleSearchTermChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setSearchTerm(e.target.value);
-        setShowCourseList(e.target.value.trim().length > 0);
-    };
-
-    const semesterOptions = semestersData.map(semester => ({
-        value: semester.id,
-        label: semester.season + " " + semester.year
-    }));
-    const [selectedSemester, setSelectedSemester] = useState<string | null>(
-        null
-    );
-
-    const handleCourseSelected = (selectedCourse: Course) => {
-        if (selectedSemester) {
-            const semesterIndex = semestersData.findIndex(
-                semester => semester.id === selectedSemester
-            );
-
-            if (semesterIndex !== -1) {
-                handleInsertCourse(semesterIndex, selectedCourse);
-            } else {
-                alert("Invalid semester selection. Course not added.");
-            }
-        } else {
-            alert("Please select a semester before adding a course.");
-        }
-    };
-
-    const handleSkipToggle = (index: number) => {
-        const updatedSemesters = [...semestersData];
-        updatedSemesters[index].skip = !updatedSemesters[index].skip;
-        setSemestersData(updatedSemesters);
-    };
-
-    const handleClearCourses = (index: number) => {
-        const updatedSemesters = [...semestersData];
-        updatedSemesters[index].courses = [];
-        setSemestersData(updatedSemesters);
-    };
-
-    const handleInsertCourse = (index: number, newCourse: Course) => {
-        const updatedSemesters = [...semestersData];
-        updatedSemesters[index].courses.push(newCourse);
-        setSemestersData(updatedSemesters);
-    };
-
-    const handleRemoveCourse = (index: number, courseIndex: number) => {
-        const updatedSemesters = [...semestersData];
-        updatedSemesters[index].courses.splice(courseIndex, 1);
-        setSemestersData(updatedSemesters);
-    };
-    const handleEditCourse = (
-        semesterIndex: number,
-        courseIndex: number,
-        updatedCourse: Course
-    ) => {
-        const updatedSemesters = [...semestersData];
-        updatedSemesters[semesterIndex].courses[courseIndex] = updatedCourse;
-        setSemestersData(updatedSemesters);
-    };
-    const handleInsertSemester = () => {
-        const lastSemester = semestersData[semestersData.length - 1];
-        const newSemester: Semester = {
-            id: getNextSemesterName(lastSemester),
-            season: getNextSeason(lastSemester),
-            year: getNextYear(lastSemester),
-            courses: [],
-            skip: false
-        };
-
-        setSemestersData(prevSemesters => [...prevSemesters, newSemester]);
-        setCoursesVisibility(prevVisibility => [...prevVisibility, true]);
-    };
-
-    const getNextSemesterName = (lastSemester: Semester | undefined) => {
-        if (!lastSemester) return "Fall 2023";
-
-        return lastSemester.season === "Fall"
-            ? `Spring ${lastSemester.year}`
-            : `Fall ${parseInt(lastSemester.year, 10) + 1}`;
-    };
-
-    const getNextSeason = (lastSemester: Semester | undefined) => {
-        if (!lastSemester) return "Fall";
-
-        return lastSemester.season === "Fall" ? "Spring" : "Fall";
-    };
-
-    const getNextYear = (lastSemester: Semester | undefined) => {
-        if (!lastSemester) return "2023";
-
-        return lastSemester.season === "Fall"
-            ? lastSemester.year
-            : (parseInt(lastSemester.year, 10) + 1).toString();
-    };
-
-    const handleRemoveSemester = (index: number) => {
-        const updatedSemesters = [...semestersData];
-        updatedSemesters.splice(index, 1);
-        setSemestersData(updatedSemesters);
-    };
-    const handleClearAllSemesters = () => {
-        setSemestersData([]);
-    };
-    const handleToggleVisibility = (index: number) => {
-        const updatedVisibility = [...coursesVisibility];
-        updatedVisibility[index] = !updatedVisibility[index];
-        setCoursesVisibility(updatedVisibility);
-    };
-    const handleToggleAllSemestersVisibility = () => {
-        setAllSemestersVisible(!allSemestersVisible);
-        setCoursesVisibility(
-            Array(semestersData.length).fill(!allSemestersVisible)
-        );
-    };
     return (
         <div className="semester-courses">
-            <select
-                value={selectedSemester || ""}
-                onChange={e => setSelectedSemester(e.target.value || null)}
-            >
-                <option value="" disabled>
-                    Selected Semester
-                </option>
-                {semesterOptions.map(option => (
-                    <option key={option.value} value={option.value}>
-                        {option.label}
-                    </option>
-                ))}
-            </select>
-            <input
-                type="text"
-                placeholder="Search by Course Code"
-                value={searchTerm}
-                onChange={handleSearchTermChange}
-            />
-            {showCourseList && (
-                <CourseList
-                    courses={filteredCourses}
-                    onCourseSelected={handleCourseSelected}
-                />
-            )}
+            <div className="plan-buttons">
+                <select
+                    value={currentPlanIndex}
+                    onChange={e => handlePlanChange(Number(e.target.value))}
+                >
+                    {plans.map((p, index) => (
+                        <option key={p.id} value={index}>
+                            {p.title}
+                        </option>
+                    ))}
+                </select>
+                <button onClick={handleInsertPlan}>Insert New Plan</button>
+                <button onClick={handleRemovePlan}>Remove Current Plan</button>
+            </div>
             <h1 onClick={handleToggleAllSemestersVisibility}>{plan.title}</h1>
             {allSemestersVisible && (
-                <>
+                <div className="button-row">
                     <button onClick={handleInsertSemester}>
                         Insert New Semester
                     </button>
                     <button onClick={handleClearAllSemesters}>
                         Clear All Semesters
                     </button>
-                </>
+                </div>
+            )}
+            <div className="search-bar">
+                <select
+                    value={selectedSemester || ""}
+                    onChange={e => setSelectedSemester(e.target.value || null)}
+                >
+                    <option value="" disabled>
+                        Selected Semester
+                    </option>
+                    {semesterOptions.map(option => (
+                        <option key={option.value} value={option.value}>
+                            {option.label}
+                        </option>
+                    ))}
+                </select>
+                <input
+                    type="text"
+                    placeholder="Search by Course Code"
+                    value={searchTerm}
+                    onChange={handleSearchTermChange}
+                />
+            </div>
+            {showCourseList && (
+                <div className="course-list-container">
+                    <CourseList
+                        courses={filteredCourses}
+                        onCourseSelected={handleCourseSelected}
+                    />
+                </div>
             )}
             {semestersData.map((semester, semesterIndex) => (
                 <div
@@ -271,6 +165,7 @@ const Planner: React.FC<Planner> = ({ plan }) => {
                                     <th>Restrictions</th>
                                     <th>Breadth</th>
                                     <th>Typically Offered</th>
+                                    <th>Actions</th>
                                 </tr>
                             </thead>
                             <tbody>
@@ -354,6 +249,71 @@ const Planner: React.FC<Planner> = ({ plan }) => {
                                             >
                                                 Remove Course
                                             </button>
+                                            <button
+                                                onClick={() =>
+                                                    handleMoveCourseStart(
+                                                        semesterIndex,
+                                                        courseIndex
+                                                    )
+                                                }
+                                            >
+                                                Move Course
+                                            </button>
+                                            {moveCourseData.fromSemesterIndex ===
+                                                semesterIndex &&
+                                                moveCourseData.courseIndex ===
+                                                    courseIndex && (
+                                                    <div>
+                                                        <select
+                                                            value={
+                                                                moveCourseData.toSemesterId
+                                                            }
+                                                            onChange={e =>
+                                                                setMoveCourseData(
+                                                                    {
+                                                                        ...moveCourseData,
+                                                                        toSemesterId:
+                                                                            e
+                                                                                .target
+                                                                                .value
+                                                                    }
+                                                                )
+                                                            }
+                                                        >
+                                                            <option
+                                                                value=""
+                                                                disabled
+                                                            >
+                                                                Select
+                                                                Destination
+                                                                Semester
+                                                            </option>
+                                                            {semesterOptions.map(
+                                                                option => (
+                                                                    <option
+                                                                        key={
+                                                                            option.value
+                                                                        }
+                                                                        value={
+                                                                            option.value
+                                                                        }
+                                                                    >
+                                                                        {
+                                                                            option.label
+                                                                        }
+                                                                    </option>
+                                                                )
+                                                            )}
+                                                        </select>
+                                                        <button
+                                                            onClick={
+                                                                handleMoveCourseConfirm
+                                                            }
+                                                        >
+                                                            Move
+                                                        </button>
+                                                    </div>
+                                                )}
                                         </td>
                                     </tr>
                                 ))}
@@ -367,4 +327,3 @@ const Planner: React.FC<Planner> = ({ plan }) => {
 };
 
 export { Planner };
-export type { Course };
