+++ src/degrees/CompSci.ts
@@ -1,0 +1,507 @@
+export const CompSciPlan = {
+    id: "123",
+    title: "Computer Science Degree",
+    semesters: [
+        {
+            id: "Fall 2022",
+            season: "Fall",
+            year: "2022",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 108",
+                    name: "Introduction to Computer Science I",
+                    descr: "Computing and principles of programming with an emphasis on systematic program design. Topics include functional programming, data abstraction, procedural abstraction, use of control and state, recursion, testing, and object-oriented programming concepts. Requires no prior programming experience, open to any major, but intended primarily for majors and minors in computer science or mathematics.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "EGGG 101",
+                    name: "Introduction to Engineering",
+                    descr: "Introduction to profession, including disciplines of chemical, civil, computer, electrical, environmental, and mechanical engineering.  Prepares students for success through integration of:  technical problem solving and engineering design, ethical decision-making, teamwork, and communicating to diverse audiences.",
+                    credits: " 2",
+                    preReq: "",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall"
+                },
+                {
+                    code: "ENGL 110",
+                    name: "Seminar in Composition",
+                    descr: "",
+                    credits: " 3",
+                    preReq: "",
+                    restrict:
+                        "Required of all students. Should be taken first year on campus.",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall, Winter, Spring, Summer"
+                },
+                {
+                    code: "MATH 241",
+                    name: "Analytic Geometry and Calculus A",
+                    descr: "Functions, limits, continuity, derivatives. Polynomial, rational, exponential, hyperbolic, logarithmic, trigonometric and inverse trigonometric functions. Definite and indefinite integrals and the Fundamental Theorem of Calculus. Simple differential equations (separable ODE, linear ODE). ODE models leading to exponential growth and decay.",
+                    credits: " 4",
+                    preReq: "MATH 117 or acceptable score on the Math Placement Exam in accordance with current standards determined by the Department of Mathematical Sciences. See https://www.mathsci.udel.edu/courses-placement/ud-math-placement for more information.",
+                    restrict:
+                        "Students who received credit in MATH 242 or MATH 243 are not eligible to take this course without permission.",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall, Winter, Spring, Summer"
+                },
+                {
+                    code: "UNIV 166",
+                    name: "Special Problem",
+                    descr: "",
+                    credits: " 1-4",
+                    preReq: "",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Verify offering with Dept"
+                }
+            ]
+        },
+        {
+            id: "Spring 2023",
+            season: "Spring",
+            year: "2023",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 181",
+                    name: "Introduction to Computer Science II",
+                    descr: "Principles of computer science illustrated and applied through programming in an object oriented language. Programming projects illustrate computational problems, styles and issues that arise in computer systems development and in all application areas of computation.",
+                    credits: " 3",
+                    preReq: "Grade of C- or better in CISC 108 or CISC 106.",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall, Summer and Spring"
+                },
+                {
+                    code: "CISC 210",
+                    name: "Introduction to Systems Programming",
+                    descr: "",
+                    credits: " 3",
+                    preReq: "A grade of C- or better in CISC 106 or CISC 108.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "MATH 210",
+                    name: "Discrete Mathematics I",
+                    descr: "Elements of sets and logic. Relations, functions. Integers. Induction and recursion. Principles and techniques of counting. Graphs. Paths and circuits",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall, Winter and Spring"
+                },
+                {
+                    code: "UNIV 167",
+                    name: "SEMINAR",
+                    descr: "",
+                    credits: " 0-6",
+                    preReq: "",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Summer"
+                },
+                {
+                    code: "GEOL 107",
+                    name: "Geology of Dynamic Earth",
+                    descr: "Principles of physical geology and its application in interpreting earth processes. Laboratory covers identification of earth materials and the interpretation of topographic and geologic maps.",
+                    credits: " 4",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GRP D LAB: A&S Math, Nat Sci & Tech w Lab",
+                    typ: "Fall"
+                }
+            ]
+        },
+        {
+            id: "Fall 2023",
+            season: "Fall",
+            year: "2023",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 220",
+                    name: "Data Structures",
+                    descr: "Review of data type abstraction, recursion, arrays, stacks, queues, multiple stacks and linked lists. Emphasis on dynamic storage management, garbage collection, trees, graphs, tables, sorting and searching.",
+                    credits: " 3",
+                    preReq: "A minimum grade of C- in CISC 210.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall, Summer and Spring"
+                },
+                {
+                    code: "CISC 260",
+                    name: "Machine Organization and Assembly Language",
+                    descr: "Introduction to the basics of machine organization. Programming tools and techniques at the machine and assembly levels. Assembly language programming and computer arithmetic techniques.",
+                    credits: " 3",
+                    preReq: "A minimum grade of C- in CISC 210.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "PHIL 100",
+                    name: "Philosophies of Life",
+                    descr: "Survey of selected past and present philosophies that people strive to live by, typically including individualistic, group-oriented, religious and nonreligious positions.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Creative Arts and Humanities; A&S: GROUP A: A&S Creative Arts & Humanities",
+                    typ: "Fall, Summer and Spring"
+                },
+                {
+                    code: "HIST 103",
+                    name: "World History to 1400",
+                    descr: "The course explores principal political, economic, cultural and social developments in world history as they relate to the present. Start and end dates (along with geographic emphasis on Asia, Africa, the Americas and Europe) vary by instructor.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: History and Cultural Change; A&S: GROUP B: A&S History & Cultural Change",
+                    typ: "Fall, Winter, Spring, Summer"
+                },
+                {
+                    code: "GEOL 105",
+                    name: "Geological Hazards and Their Human Impact",
+                    descr: "Geological processes and events that adversely affect humans and civilization. Methods for predicting and dealing with geological hazards. May be taken with a 1 credit lab, GEOL 115.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall, Summer and Spring"
+                },
+                {
+                    code: "GEOL 115",
+                    name: "Geological Hazards Laboratory",
+                    descr: "Laboratory investigation of environmental geological processes and their impact on society: earthquakes, volcanoes, coastal hazards, rivers and streams, groundwater. Identification of earth materials and study of earth processes using topographical, geological and tectonic maps.",
+                    credits: " 1",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GRP D LAB: A&S Math, Nat Sci & Tech w Lab",
+                    typ: "Fall, Summer and Spring"
+                }
+            ]
+        },
+        {
+            id: "Spring 2024",
+            season: "Spring",
+            year: "2024",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 275",
+                    name: "Introduction to Software Engineering",
+                    descr: "Object oriented software design and development through use of an object oriented programming language. Topics include team programming, design patterns, graphical user interfaces, software engineering tools (e.g., integrated development environments, version control, build management, bug tracking, automated testing).",
+                    credits: " 3",
+                    preReq: "Minimum grade of C- in CISC 181 and CISC 220.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "CISC 355",
+                    name: "Computers, Ethics and Society",
+                    descr: "",
+                    credits: " 3",
+                    preReq: "",
+                    restrict:
+                        "Cannot be used as a technical elective for CISC or INSY majors or minors.",
+                    breadth:
+                        "University: Creative Arts and Humanities; A&S: GROUP A: A&S Creative Arts & Humanities",
+                    typ: "Fall, Winter and Spring"
+                },
+                {
+                    code: "ENWC 201",
+                    name: "Wildlife Conservation and Ecology",
+                    descr: "",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "Should precede BISC 302.",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "SOCI 201",
+                    name: "Introduction to Sociology",
+                    descr: "An overview of the sociological perspective of the study of society, social organization and social institutions with special emphasis on the social causes and consequences of human behavior.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Social and Behavioral Sciences; A&S: GROUP C: A&S Social & Behavioral Sci",
+                    typ: "Fall, Summer and Spring"
+                },
+                {
+                    code: "CISC 320",
+                    name: "Introduction to Algorithms",
+                    descr: "Design and analysis of algorithms: worst/average case analysis, proofs for correctness and performance of algorithms. Algorithmic strategies (divide and conquer, greedy methods, dynamic programming, etc.). Algorithms for searching, forming and traversal of strings, trees and graphs. Categorization of computational problems: classes P and NP. NP completeness.",
+                    credits: " 3",
+                    preReq: "MATH 210 and a minimum grade of C- in CISC 220.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                }
+            ]
+        },
+        {
+            id: "Fall 2024",
+            season: "Fall",
+            year: "2024",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 303",
+                    name: "Automata Theory",
+                    descr: "Automata and formal language theory. Background for advanced applications in compilers, computer networks, operating systems, and natural language processing. Finite automata and regular languages. Push down automata and context free grammars. Turing machines.",
+                    credits: " 3",
+                    preReq: "A minimum grade of C- in both MATH 210 and CISC 220.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "ENGL 410",
+                    name: "Technical Writing",
+                    descr: "",
+                    credits: " 3",
+                    preReq: "ENGL 110.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall, Winter and Spring"
+                },
+                {
+                    code: "CISC 437",
+                    name: "Database Systems",
+                    descr: "Physical and logical organization of databases. Data retrieval languages, relational database languages, security and integrity, concurrency, distributed databases.",
+                    credits: " 3",
+                    preReq: "A minimum grade of C- in CISC 220.",
+                    restrict:
+                        "Students who received credit in CISC637 are not eligible to take this course without permission.",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "PSYC 100",
+                    name: "General Psychology",
+                    descr: "Introduction to the process of psychological science. Includes coverage of research methods, biological bases of behavior, sensation and perception, cognitive psychology, abnormal behavior and treatment, developmental psychology, and social and personality psychology.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Social and Behavioral Sciences; A&S: GROUP C: A&S Social & Behavioral Sci",
+                    typ: "Fall, Summer and Spring"
+                },
+                {
+                    code: "NTDT 200",
+                    name: "Nutrition Concepts",
+                    descr: "Functions and sources of nutrients, dietary adequacy, energy balance and metabolism with emphasis on health promotion. Includes weight control, evaluation of popular diets, vegetarianism, eating disorders, alcohol, other current issues and concerns in nutrition.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall, Summer and Spring"
+                }
+            ]
+        },
+        {
+            id: "Spring 2025",
+            season: "Spring",
+            year: "2025",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 304",
+                    name: "Logic for Programming",
+                    descr: "Propositional and predicate logic for general reasoning and advanced applications in knowledge representation in artificial intelligence and database, program correctness and programming semantics. Models, resolution, logic programming, and natural deduction.",
+                    credits: " 3",
+                    preReq: "CISC 220, MATH 210 (with minimum C- grade in both).",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Spring"
+                },
+                {
+                    code: "CISC 361",
+                    name: "Operating Systems",
+                    descr: "Principles and techniques employed in the development of operating systems and their control programs. Includes management of memory, processors, I/O devices.",
+                    credits: " 3",
+                    preReq: "A minimum grade of C- in both CISC 220 and CISC 260. CPEG 222 may be substituted for CISC 260.",
+                    restrict: "CISC 360 recommended as a prerequisite.",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "SPAN 105",
+                    name: "Spanish I - Elementary",
+                    descr: "Introduction to the Spanish language and a development of speaking, listening, reading and writing skills through the use of basic texts.",
+                    credits: " 4",
+                    preReq: "",
+                    restrict:
+                        "No Spanish background, two or fewer years of high school Spanish. .",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall, Winter, Spring, Summer"
+                },
+                {
+                    code: "ANTH 205",
+                    name: "Anthropology and Human Nature",
+                    descr: "Archaeological, biological and cultural anthropology organized around the theme of human nature.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Creative Arts and Humanities; A&S: GROUP A: A&S Creative Arts & Humanities",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "WOMS 201",
+                    name: "Introduction to Women and Gender Studies",
+                    descr: "Study of causes and conditions determining women\u2019s status in society, as evidenced in institutional structures and personal relations between men and women. Taught from multidisciplinary perspectives.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Social and Behavioral Sciences; A&S: GROUP C: A&S Social & Behavioral Sci",
+                    typ: "Fall, Winter, Spring, Summer"
+                }
+            ]
+        },
+        {
+            id: "Fall 2025",
+            season: "Fall",
+            year: "2025",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 481",
+                    name: "Artificial Intelligence",
+                    descr: "Programming techniques for problems not amenable to algorithmic solutions. Problem formulation, search strategies, state spaces, applications of logic, knowledge representation, planning and application areas.",
+                    credits: " 3",
+                    preReq: "CISC 220 with a minimum grade of C-, CISC 304.",
+                    restrict:
+                        "Students who received credit in CISC681 are not eligible to take this course without permission.",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                },
+                {
+                    code: "LLCU 338",
+                    name: "Light and Shadow\u2013Japanese Films",
+                    descr: "Inquiry into Japanese films from the immediate postwar period to present.Both aesthetic and technical elements of film production and receptionfeatured. Discussion intensive.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Creative Arts and Humanities; A&S: GROUP A: A&S Creative Arts & Humanities",
+                    typ: "Verify offering with Dept"
+                },
+                {
+                    code: "CISC 498",
+                    name: "Computer Science Senior Design Project I",
+                    descr: "First semester of two-semester senior software design course.  Students work in teams to develop solutions to real-world problems for a client.  Application of a modern software engineering process; requirement solicitation, analysis, and specification; prototyping; design; incremental development; testing and verification; client interaction; and presentation. Offered in Fall only.",
+                    credits: " 3",
+                    preReq: "CISC 275 and CISC 320.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall"
+                },
+                {
+                    code: "HIST 104",
+                    name: "World History since 1400",
+                    descr: "The course explores principal political, economic, cultural and social developments in world history as they relate to the present. Start and end dates (along with geographic emphasis on Asia, Africa, the Americas and Europe) vary by instructor.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: History and Cultural Change; A&S: GROUP B: A&S History & Cultural Change",
+                    typ: "Fall, Winter, Spring, Summer"
+                },
+                {
+                    code: "PHYS 145",
+                    name: "Black Holes and Cosmic Evolution",
+                    descr: "Geared toward a general audience and focuses on topics such as finding black holes in space; the nature of space, time and black holes; the origin of the Universe; and the development of complex life here on Earth.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict: "",
+                    breadth:
+                        "University: Mathematics, Natural Sciences and Technology; A&S: GROUP D: A&S Math, Nat Sci & Technology",
+                    typ: "Fall and Spring"
+                }
+            ]
+        },
+        {
+            id: "Spring 2026",
+            season: "Spring",
+            year: "2026",
+            skip: false,
+            courses: [
+                {
+                    code: "CISC 499",
+                    name: "Computer Science Senior Design Project II",
+                    descr: "",
+                    credits: " 3",
+                    preReq: "CISC 498.",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Spring"
+                },
+                {
+                    code: "CISC 484",
+                    name: "Introduction to Machine Learning",
+                    descr: "Development of methods to learn to solve a task using examples. Explore different machine learning algorithms/techniques and discuss their strengths and weaknesses and situations they are or are not suited for.",
+                    credits: " 3",
+                    preReq: "CISC 220 and MATH 205 or MATH 350.",
+                    restrict:
+                        "Students who received credit in CISC684 are not eligible to take this course without permission.",
+                    breadth: "University: ; A&S: ",
+                    typ: "Spring"
+                },
+                {
+                    code: "CISC 489",
+                    name: "Topics: Artificial Intelligence",
+                    descr: "Contents vary to coincide with the interests of students and current faculty.",
+                    credits: " 3",
+                    preReq: "",
+                    restrict:
+                        "Students who received credit in CISC689 are not eligible to take this course without permission.",
+                    breadth: "University: ; A&S: ",
+                    typ: "Verify offering with Dept"
+                },
+                {
+                    code: "UNIV 367",
+                    name: "SEMINAR",
+                    descr: "",
+                    credits: " 0-6",
+                    preReq: "",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Spring"
+                },
+                {
+                    code: "UNIV 114",
+                    name: "Skilmod: Critical Thinking",
+                    descr: "Teaches strategies of good reasoning in humanities, social sciences and other liberal arts areas. Students learn how to become independent, well-informed thinkers and how to make persuasive arguments, using their own course materials.",
+                    credits: " 1",
+                    preReq: "",
+                    restrict: "",
+                    breadth: "University: ; A&S: ",
+                    typ: "Fall and Spring"
+                }
+            ]
+        }
+    ]
+};
