+++ src/degrees/EmptyPlan.ts
@@ -1,0 +1,62 @@
+export const EmptyPlan = {
+    id: "123",
+    title: "Computer Science Degree",
+    semesters: [
+        {
+            id: "Fall 2022",
+            season: "Fall",
+            year: "2022",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Spring 2023",
+            season: "Spring",
+            year: "2023",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Fall 2023",
+            season: "Fall",
+            year: "2023",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Spring 2024",
+            season: "Spring",
+            year: "2024",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Fall 2024",
+            season: "Fall",
+            year: "2024",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Spring 2025",
+            season: "Spring",
+            year: "2025",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Fall 2025",
+            season: "Fall",
+            year: "2025",
+            skip: false,
+            courses: []
+        },
+        {
+            id: "Spring 2026",
+            season: "Spring",
+            year: "2026",
+            skip: false,
+            courses: []
+        }
+    ]
+};
