+++ src/interfaces/Course.ts
@@ -1,0 +1,10 @@
+export interface Course {
+    code: string;
+    name: string;
+    descr: string;
+    credits: string;
+    preReq: string;
+    restrict: string;
+    breadth: string;
+    typ: string;
+}
