+++ src/interfaces/Plan.ts
@@ -1,0 +1,7 @@
+import { Semester } from "./Semester";
+
+export interface Plan {
+    id: string;
+    title: string;
+    semesters: Semester[];
+}
