+++ src/interfaces/Semester.ts
@@ -1,0 +1,9 @@
+import { Course } from "./Course";
+
+export interface Semester {
+    id: string;
+    season: string;
+    year: string;
+    courses: Course[];
+    skip: boolean;
+}
